requirements.txt
Flask==2.2.5
langchain==0.3.24
langchain-core==0.3.55
langchain-community==0.3.22
langchain-openai==0.3.14
langchain-chroma==0.2.3
python-dotenv==0.21.0
chromadb==0.4.22
pypdf==4.0.2
openai>=1.68.2,<2.0.0
tiktoken>=0.7,<1.0.0

